requirements.txt
gradio==3.41.2
duckduckgo_search==7.3.0
requests==2.32.3
huggingface-hub==0.28.1
langchain==0.3.17
langchain-community==0.3.16
langgraph==0.2.68
setuptools==75.8.0
aws-cdk-lib==2.177.0
fastapi==0.115.8
uvicorn==0.34.0